ActionsDropdown and Underline shouldn’t be Text decorators
Decorators shouldn’t alter the structure of the
wrapped component, i.e. wrap it with another
container.
