+++ app/src/shared/components/ActionsDropdown.jsx
@@ -1,0 +1,56 @@
+// @flow
+
+import React, { type Node, useState } from 'react'
+import styled, { css } from 'styled-components'
+import Meatball from '$shared/components/Meatball'
+import DropdownActions from '$shared/components/DropdownActions'
+import { type UseStateTuple } from '$shared/flowtype/common-types'
+
+const ActionContainer = styled.div`
+    display: inline-block;
+    position: absolute;
+    right: 0.5rem;
+    top: 50%;
+    transform: translate(0, -50%);
+
+    ${({ open }) => !!open && css`
+        z-index: 2;
+    `}
+`
+
+type Props = {
+    actions?: Array<typeof DropdownActions.Item>,
+    children?: Node,
+}
+
+const UnstyledActionsDropdown = ({ actions, children = null, ...props }: Props) => {
+    const [open, setOpen]: UseStateTuple<boolean> = useState(false)
+
+    return !actions || !actions.length ? (
+        children
+    ) : (
+        <div {...props}>
+            {children}
+            <ActionContainer open={open}>
+                <DropdownActions
+                    title={(
+                        <Meatball alt="Actions" gray />
+                    )}
+                    menuProps={{
+                        right: true,
+                    }}
+                    onMenuToggle={setOpen}
+                    noCaret
+                >
+                    {actions}
+                </DropdownActions>
+            </ActionContainer>
+        </div>
+    )
+}
+
+const ActionsDropdown = styled(UnstyledActionsDropdown)`
+    position: relative;
+`
+
+export default ActionsDropdown
