+++ app/src/shared/components/Input/Text/index.jsx
@@ -5,10 +5,8 @@ import { compose } from 'redux'
 import OnAutoCompleteDecorator from './OnAutoCompleteDecorator'
 import OnCommitDecorator from './OnCommitDecorator'
 import StatefulInputDecorator from './StatefulInputDecorator'
-import ActionsDropdownDecorator from './ActionsDropdownDecorator'
 import FlushHistoryDecorator from './FlushHistoryDecorator'
 import SelectAllOnFocusDecorator from './SelectAllOnFocusDecorator'
-import UnderlineDecorator from './UnderlineDecorator'
 
 type Props = {
     tag: 'input' | 'textarea',
@@ -19,11 +17,9 @@ const Input = ({ tag: Tag = 'input', ...props }: Props) => (
 )
 
 export default compose(
-    ActionsDropdownDecorator,
     FlushHistoryDecorator,
     OnAutoCompleteDecorator,
     OnCommitDecorator,
     SelectAllOnFocusDecorator,
     StatefulInputDecorator,
-    UnderlineDecorator,
 )(Input)
